fix(cnn): 修复并优化CNN模块的多项配置校验、BN计算及梯度问题
- 修正CNN推理和训练代码的配置校验逻辑：将原有的配置大小下限检查改为严格匹配结构体大小，新增配置类型名称校验，避免误用错误的配置结构体
- 调整CnnTrainConfig中batch_size字段的类型为uint32_t，统一类型定义
- 优化随机权重生成的归一化计算，使用完整的32位无符号整数范围提升随机数分布均匀性
- 将BN计算中硬编码的1e-8阈值替换为配置项中的bn_epsilon，提升模块灵活性
- 修正tanh激活函数的导数计算表达式，优化浮点计算稳定性
- 重写CNN训练反向传播的BN梯度计算逻辑，实现完整的批量归一化反向传播公式，提升训练精度

+++ src/nn/types/cnn/cnn_config.h
@@ -76,7 +76,7 @@ typedef struct {
     float learning_rate;                  /**< Step size used by the simple SGD update. */
     float momentum;                       /**< Reserved for future optimizer growth. */
     float weight_decay;                   /**< L2-style decay applied to trainable weights. */
-    size_t batch_size;                    /**< Batch size requested by generated wrappers. */
+    uint32_t batch_size;                  /**< Batch size requested by generated wrappers. */
     uint32_t seed;                        /**< Reserved deterministic seed for future train state. */
 } CnnTrainConfig;
